pyproject.toml
[project]
name = "post-code-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "gitpython>=3.1.44",
    "google-ai-generativelanguage>=0.6.17",
    "google-genai>=1.8.0",
    "pydantic>=2.11.1",
]


[project.scripts]
post-code-summary = "pre_commit_hooks.post_code_summary:main"
